+++ R/proxy.R
@@ -38,6 +38,7 @@
 #'   \item `ejaccard`: the real value version of `jaccard`
 #'   \item `dice`: Dice coefficient
 #'   \item `edice`: the real value version of `dice`
+#'   \item `hamann`: Hamann similarity
 #'   \item `faith`: Faith similarity
 #'   \item `simple matching`: the percentage of common elements
 #' }
@@ -104,14 +105,15 @@ proxy <- function(x, y = NULL, margin = 1,
 
     method[method == "hamman"] <- "hamann" # for transition
     method <- match.arg(method)
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
-
+    #x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(as(x, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     symm <- is.null(y)
 
     if (is.null(y)) {
         y <- x
     } else {
-        y <- as(as(y, "CsparseMatrix"), "dgCMatrix")
+        #y <- as(as(y, "CsparseMatrix"), "dgCMatrix")
+        y <- as(as(as(y, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     }
     if (!margin %in% c(1, 2))
         stop("Matrgin must be 1 (row) or 2 (column)")